+++ packages/overture-schema-transportation-theme/src/overture/schema/transportation/models.py
@@ -8,6 +8,7 @@ from overture.schema.core import (
     StrictBaseModel,
 )
 from overture.schema.core.models import GeometricRangeScope
+from overture.schema.core.primitives import float64, int32
 from overture.schema.core.types import (
     Id,
     Level,
@@ -44,10 +45,10 @@ from .enums import (
 )
 
 SpeedValue = NewType(
-    "SpeedValue", Annotated[int, Field(ge=1, le=350, description="Speed value")]
+    "SpeedValue", Annotated[int32, Field(ge=1, le=350, description="Speed value")]
 )
 
-Width = NewType("Width", Annotated[float, Field(gt=0)])
+Width = NewType("Width", Annotated[float64, Field(gt=0)])
 
 
 class ConnectorReference(StrictBaseModel):
@@ -179,23 +180,23 @@ class Speed(StrictBaseModel):
 
 
 class IsMoreThanIntegerRelation(StrictBaseModel):
-    is_more_than: int
+    is_more_than: int32
 
 
 class IsAtLeastIntegerRelation(StrictBaseModel):
-    is_at_least: int
+    is_at_least: int32
 
 
 class IsEqualToIntegerRelation(StrictBaseModel):
-    is_equal_to: int
+    is_equal_to: int32
 
 
 class IsAtMostIntegerRelation(StrictBaseModel):
-    is_at_most: int
+    is_at_most: int32
 
 
 class IsLessThanIntegerRelation(StrictBaseModel):
-    is_less_than: int
+    is_less_than: int32
 
 
 IntegerRelation = Annotated[
@@ -216,7 +217,7 @@ class LengthValueWithUnit(StrictBaseModel):
     # Required
 
     unit: LengthUnit
-    value: Annotated[float, Field(ge=0)]
+    value: Annotated[float64, Field(ge=0)]
 
 
 class IsMoreThanLengthRelation(StrictBaseModel):
@@ -257,7 +258,7 @@ class WeightValueWithUnit(StrictBaseModel):
     # Required
 
     unit: WeightUnit
-    value: Annotated[float, Field(ge=0)]
+    value: Annotated[float64, Field(ge=0)]
 
 
 class IsMoreThanWeightRelation(StrictBaseModel):
@@ -383,7 +384,7 @@ class VehicleScopeRule(StrictBaseModel):
 
     dimension: VehicleDimension
     comparison: VehicleComparison
-    value: Annotated[float, Field(ge=0)]
+    value: Annotated[float64, Field(ge=0)]
 
     # Optional
